+++ BuildingBlocks/src/BuildingBlocks.Application/Identities/AggregateType.cs
@@ -1,0 +1,17 @@
+﻿namespace Backbone.BuildingBlocks.Application.Identities;
+public enum AggregateType
+{
+    Challenges,
+    PnsRegistrations,
+    Identities,
+    Files,
+    Messages,
+    QuotaIdentities,
+    Relationships,
+    RelationshipTemplates,
+    RelationshipTemplateAllocations,
+    ExternalEvents,
+    SyncRuns,
+    Datawallets,
+    Tokens
+}
