+++ BuildingBlocks/src/BuildingBlocks.Application/Identities/IDeletionProcessLogger.cs
@@ -1,0 +1,7 @@
+﻿using Backbone.DevelopmentKit.Identity.ValueObjects;
+
+namespace Backbone.BuildingBlocks.Application.Identities;
+public interface IDeletionProcessLogger
+{
+    Task LogDeletion(IdentityAddress identityAddress, AggregateType aggregateType);
+}
