+++ BuildingBlocks/src/BuildingBlocks.Application/Identities/IIdentityDeleter.cs
@@ -5,25 +5,3 @@ public interface IIdentityDeleter
 {
     Task Delete(IdentityAddress identityAddress, IDeletionProcessLogger deletionProcessLogger);
 }
-
-public interface IDeletionProcessLogger
-{
-    Task LogDeletion(IdentityAddress identityAddress, AggregateType aggregateType);
-}
-
-public enum AggregateType
-{
-    Challenges,
-    PnsRegistrations,
-    Identities,
-    Files,
-    Messages,
-    QuotaIdentities,
-    Relationships,
-    RelationshipTemplates,
-    RelationshipTemplateAllocations,
-    ExternalEvents,
-    SyncRuns,
-    Datawallets,
-    Tokens
-}
